backend/requirements.txt
Flask==3.0.3
mysql-connector-python==9.3.0
flask-cors==5.0.1
python-dotenv==1.0.1
spotipy==2.25.1